:truck: Re-standardize, moving action to beginning of names.
Why? There is only one type my vs. non-my components, but there are potentially more action types.
e.g. read, view, manage.
