+++ web/src/modules/travel-authorizations/components/edit-my-travel-authorization-details-page/ApprovalsFormCard.vue
@@ -20,7 +20,7 @@
             <v-btn
               v-if="!refreshingEstimatesSilently && hasEstimates"
               :to="{
-                name: 'MyTravelAuthorizationEditEstimatePage',
+                name: 'EditMyTravelAuthorizationEstimatePage',
                 params: { formId: currentTravelAuthorizationId },
               }"
               class="mt-1"
@@ -102,7 +102,7 @@ import { mapActions, mapState, mapGetters } from "vuex"
 import preApprovedTravelRequestsApi from "@/api/pre-approved-travel-requests-api"
 
 import SearchableUserEmailCombobox from "@/components/SearchableUserEmailCombobox"
-import EstimateGenerateDialog from "@/modules/travel-authorizations/components/my-travel-authorization-edit-estimate-page/EstimateGenerateDialog"
+import EstimateGenerateDialog from "@/modules/travel-authorizations/components/edit-my-travel-authorization-estimate-page/EstimateGenerateDialog"
 
 import EstimatedCostTextField from "./approvals-form-card/EstimatedCostTextField"
 import SubmitToSupervisorButton from "./approvals-form-card/SubmitToSupervisorButton"
