+++ web/src/modules/travel-authorizations/components/edit-my-travel-authorization-details-page/DetailsFormCard.vue
