+++ web/src/modules/travel-authorizations/components/edit-my-travel-authorization-details-page/details-form-card/MultiDestinationStopsSection.vue
