+++ web/src/modules/travel-authorizations/components/read-my-travel-authorization-details-page/PurposeCard.vue
