+++ web/src/modules/travel-authorizations/components/read-my-travel-authorization-details-page/details-card/OneWayStopsSection.vue
