+++ web/src/modules/travel-authorizations/components/read-my-travel-authorization-estimate-page/EstimatesTable.vue
