+++ web/src/modules/travel-authorizations/layouts/README.md
@@ -1,3 +1,4 @@
 # Travel Authorization Layouts
 
 Layouts are components that wrap a page or collection of pages, then can't be routed to directly.
+Naming convention is `<action><object>Layout`.
