+++ web/src/modules/travel-authorizations/pages/README.md
@@ -5,5 +5,5 @@ Component names should end in `Page`.
 Page specific components should be in the components/name-of-page-component directory.
 
 e.g.
-- `web/src/modules/travel-authorizations/pages/MyTravelAuthorizationEditDetailsPage.vue`
-has specific components in `web/src/modules/travel-authorizations/components/my-travel-authorization-edit-details-page`
+- `web/src/modules/travel-authorizations/pages/EditMyTravelAuthorizationDetailsPage.vue`
+has specific components in `web/src/modules/travel-authorizations/components/edit-my-travel-authorization-details-page`
